Create CONTRIBUTING.md
It's a bit difficult for the beginners to understand git. But we accept people from all diversity. For this reason it's better to have a well constructed Contributing Guideline on a open source project. Some cases may also be found for a contribution the whole project got corrupted. No one wants that. So I want to help you in this cause.